requirements.txt
python-docx
beautifulsoup4
lxml
pillow
requests
markdown
psutil
latex2mathml